Address Copilot review round 2: fold tests, async handlers, patch bump
- Fold task-delete.spec.ts into existing task-deletion.spec.ts
- Make dialog handlers async with await
- Change bump type from none to patch
