+++ common/changes/@grackle-ai/cli/fix-live-testing-bugs-followup_2026-03-10-20-44.json
@@ -2,7 +2,7 @@
   "changes": [
     {
       "comment": "Review-fixes-and-test-coverage",
-      "type": "none",
+      "type": "patch",
       "packageName": "@grackle-ai/cli"
     }
   ],
